+++ src/components/organisms/PlayerBar/index.ts
@@ -2,18 +2,18 @@ import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 
 import PlayerBar from './PlayerBar'
-import { updateVolume as updateVolumeAction } from '../../../actions/soundActions'
-import { increaseSongTime as increaseSongTimeAction } from '../../../actions/songActions'
+import { updateVolume } from '../../../features/soundSlice'
+import { increaseSongTime } from '../../../features/songsSlice'
 
-const mapStateToProps = ({ songsReducer, soundReducer }) => ({
-  songDetails: songsReducer.songDetails,
-  volume: soundReducer.volume,
+const mapStateToProps = ({ songs, sound }) => ({
+  songDetails: songs.songDetails,
+  volume: sound.volume,
 })
 
 const mapDispatchToProps = (dispatch) => {
   return bindActionCreators({
-    updateVolumeAction,
-    increaseSongTimeAction,
+    updateVolume,
+    increaseSongTime,
   }, dispatch)
 }