+++ src/constants/index.ts
@@ -1,1 +1,3 @@
 export { imageSizes } from './api'
+
+export { trackPlaybackStatus } from './track'